BE-Module-13/e-commerce_api/requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
Flask-Limiter==1.5.0
Flask-Caching
Flask-Cors==3.0.10
SQLAlchemy==1.4.19
Werkzeug==2.3.3
flask_migrate
flask_marshmallow
flask_jwt_extended
cryptography
connexion
circuitbreaker
pyjwt